Refactor answer variables from float to integer and make user_answer nullable (#30)
* Initial plan

* Refactor answer variables to integer type in backend


* Allow user_answer to be nullable for LLM failures


* fix

---------

+++ services/quiz_history/service.py
@@ -85,8 +85,8 @@ class QuizHistoryService:
                     "username": {"type": "keyword"},
                     "question": {"type": "text"},
                     "user_equation": {"type": "text"},
-                    "user_answer": {"type": "float"},
-                    "correct_answer": {"type": "float"},
+                    "user_answer": {"type": "integer"},
+                    "correct_answer": {"type": "integer"},
                     "is_correct": {"type": "boolean"},
                     "category": {"type": "keyword"},
                     "timestamp": {"type": "date"}
@@ -278,8 +278,8 @@ if __name__ == "__main__":
             username="test_user",
             question="What is 5 + 3?",
             user_equation="5 + 3",
-            user_answer=8.0,
-            correct_answer=8.0,
+            user_answer=8,
+            correct_answer=8,
             is_correct=True,
             category="addition",
             timestamp=datetime.now()
